requirements/cpu-requirements.txt
bentoml==1.3.19
Pillow==10.2.0
https://download.pytorch.org/whl/cpu/torch-2.3.0%2Bcpu-cp311-cp311-linux_x86_64.whl #torch==2.3.0+cpu
https://download.pytorch.org/whl/cpu/torchvision-0.18.0%2Bcpu-cp311-cp311-linux_x86_64.whl #torchvision==0.18.0+cpu